refactor: rename entity to model
