+++ src/main/java/kr/co/wanted/backend31/common/model/brand/Brand.java
@@ -1,4 +1,4 @@
-package kr.co.wanted.backend31.common.entity.category;
+package kr.co.wanted.backend31.common.model.brand;
 
 import java.util.List;
 
@@ -9,10 +9,9 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+import kr.co.wanted.backend31.common.model.product.Product;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,8 +20,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "categories")
-public class Category {
+@Table(name = "brands")
+public class Brand {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -34,17 +33,11 @@ public class Category {
     private String slug;
     @Column(name = "description")
     private String description;
-    @Column(name = "level")
-    private Integer level;
-    @Column(name = "image_url")
-    private String imageUrl;
+    @Column(name = "logo_url")
+    private String logoUrl;
+    @Column(name = "website")
+    private String website;
 
-    @JoinColumn(name = "parent_id")
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Category parentCategory;
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "parentCategory", cascade = CascadeType.ALL)
-    private List<Category> childrens;
-
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "category", cascade = CascadeType.ALL)
-    private List<ProductCategory> productCategories;
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "brand", cascade = CascadeType.ALL)
+    private List<Product> products;
 }
