+++ src/main/java/kr/co/wanted/backend31/common/model/product/Product.java
@@ -1,4 +1,4 @@
-package kr.co.wanted.backend31.common.entity.product;
+package kr.co.wanted.backend31.common.model.product;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -18,10 +18,10 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
-import kr.co.wanted.backend31.common.entity.brand.Brand;
-import kr.co.wanted.backend31.common.entity.category.ProductCategory;
-import kr.co.wanted.backend31.common.entity.seller.Seller;
-import kr.co.wanted.backend31.common.entity.tag.ProductTag;
+import kr.co.wanted.backend31.common.model.brand.Brand;
+import kr.co.wanted.backend31.common.model.category.ProductCategory;
+import kr.co.wanted.backend31.common.model.seller.Seller;
+import kr.co.wanted.backend31.common.model.tag.ProductTag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
