+++ src/main/java/kr/co/wanted/backend31/common/model/product/ProductOptionGroup.java
@@ -1,5 +1,8 @@
-package kr.co.wanted.backend31.common.entity.product;
+package kr.co.wanted.backend31.common.model.product;
 
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,6 +11,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -17,27 +21,23 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "product_images")
-public class ProductImage {
-    
+@Table(name = "product_option_groups")
+public class ProductOptionGroup {
+
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "url")
-    private String url;
-    @Column(name = "alt_text")
-    private String altText;
-    @Column(name = "is_primary")
-    private boolean isPrimary;
+    @Column(name = "name")
+    private String name;
     @Column(name = "display_order")
     private Integer displayOrder;
 
     @JoinColumn(name = "product_id")
     @ManyToOne(fetch = FetchType.LAZY)
     private Product product;
-    @JoinColumn(name = "option_id")
-    @ManyToOne(fetch = FetchType.LAZY)
-    private ProductOption option;
+    
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "optionGroup", cascade = CascadeType.ALL)
+    private List<ProductOption> options;
 }
