+++ src/main/java/kr/co/wanted/backend31/common/model/product/ProductPrice.java
@@ -1,8 +1,7 @@
-package kr.co.wanted.backend31.common.entity.product;
+package kr.co.wanted.backend31.common.model.product;
 
-import java.util.List;
+import java.math.BigDecimal;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -10,8 +9,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -21,23 +19,27 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "product_option_groups")
-public class ProductOptionGroup {
-
+@Table(name = "product_prices")
+public class ProductPrice {
+    
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "name")
-    private String name;
-    @Column(name = "display_order")
-    private Integer displayOrder;
+    @Column(name = "base_price")
+    private BigDecimal basePrice;
+    @Column(name = "sale_price")
+    private BigDecimal salePrice;
+    @Column(name = "cost_price")
+    private BigDecimal costPrice;
+    @Column(name = "currency")
+    private String currency;
+    @Column(name = "tax_rate")
+    private BigDecimal taxRate;
 
     @JoinColumn(name = "product_id")
-    @ManyToOne(fetch = FetchType.LAZY)
+    @OneToOne(fetch = FetchType.LAZY)
     private Product product;
-    
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "optionGroup", cascade = CascadeType.ALL)
-    private List<ProductOption> options;
+
 }
