+++ src/main/java/kr/co/wanted/backend31/common/model/product/specification/ProductCategoryCreateSpecification.java
@@ -1,9 +1,9 @@
-package kr.co.wanted.backend31.common.entity.product.specification;
+package kr.co.wanted.backend31.common.model.product.specification;
 
 import java.util.Objects;
 import java.util.function.Predicate;
 
-import kr.co.wanted.backend31.common.entity.category.ProductCategory;
+import kr.co.wanted.backend31.common.model.category.ProductCategory;
 import lombok.Builder;
 
 /**
