+++ src/main/java/kr/co/wanted/backend31/common/model/product/specification/ProductDetailCreateSpecification.java
@@ -1,11 +1,11 @@
-package kr.co.wanted.backend31.common.entity.product.specification;
+package kr.co.wanted.backend31.common.model.product.specification;
 
 import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Predicate;
 
-import kr.co.wanted.backend31.common.entity.product.ProductDetail;
+import kr.co.wanted.backend31.common.model.product.ProductDetail;
 import lombok.Builder;
 
 /**
