+++ src/main/java/kr/co/wanted/backend31/common/model/product/specification/ProductTagCreateSpecification.java
@@ -1,9 +1,9 @@
-package kr.co.wanted.backend31.common.entity.product.specification;
+package kr.co.wanted.backend31.common.model.product.specification;
 
 import java.util.Objects;
 import java.util.function.Predicate;
 
-import kr.co.wanted.backend31.common.entity.tag.ProductTag;
+import kr.co.wanted.backend31.common.model.tag.ProductTag;
 import lombok.Builder;
 
 /**
