+++ src/main/java/kr/co/wanted/backend31/common/model/tag/ProductTag.java
@@ -1,6 +1,4 @@
-package kr.co.wanted.backend31.common.entity.product;
-
-import java.math.BigDecimal;
+package kr.co.wanted.backend31.common.model.tag;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,8 +7,9 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import kr.co.wanted.backend31.common.model.product.Product;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -19,27 +18,18 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "product_prices")
-public class ProductPrice {
-    
+@Table(name = "product_tags")
+public class ProductTag {
+
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "base_price")
-    private BigDecimal basePrice;
-    @Column(name = "sale_price")
-    private BigDecimal salePrice;
-    @Column(name = "cost_price")
-    private BigDecimal costPrice;
-    @Column(name = "currency")
-    private String currency;
-    @Column(name = "tax_rate")
-    private BigDecimal taxRate;
-
     @JoinColumn(name = "product_id")
-    @OneToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     private Product product;
-
+    @JoinColumn(name = "tag_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Tag tag;
 }
