+++ src/main/java/kr/co/wanted/backend31/common/model/user/User.java
@@ -1,6 +1,9 @@
-package kr.co.wanted.backend31.common.entity.tag;
+package kr.co.wanted.backend31.common.model.user;
 
+import java.time.LocalDateTime;
 import java.util.List;
+
+import org.hibernate.annotations.CreationTimestamp;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -11,16 +14,15 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+import kr.co.wanted.backend31.common.model.product.ProductReview;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "tags")
-public class Tag {
+@Table(name = "users")
+public class User {
 
     @Id
     @Column(name = "id")
@@ -29,9 +31,14 @@ public class Tag {
 
     @Column(name = "name")
     private String name;
-    @Column(name = "slug")
-    private String slug;
+    @Column(name = "email")
+    private String email;
+    @Column(name = "avatar_url")
+    private String avatarUrl;
+    @CreationTimestamp
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
 
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "tag", cascade = CascadeType.ALL)
-    private List<ProductTag> products;
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user", cascade = CascadeType.ALL)
+    private List<ProductReview> reviews;
 }
